backend/requirements/base.txt
pytz==2020.1
psycopg2==2.8.5 --no-binary psycopg2

# Django
# ------------------------------------------------------------------------------
django==3.0.8

# django rest framework
djangorestframework==3.11.0

## kafka
kafka-python==2.0.1

## pytest
pytest_django==3.9.0
